enh: import base report-capable interfaces from niworkflows
This sidetracks #71 for interfaces with little discussion about whether
they can be consider core enough to be maintained within NiReports.

In addition, moves the SimpleBeforeAfter interface into the set of basic
interfaces, as it is very often used.
